feat: integrate RAM/index improvements and benchmark harness

+++ src/quantizer/mod.rs
@@ -4,4 +4,5 @@ pub mod mse;
 pub mod prod;
 pub mod qjl;
 
-pub type CodeIndex = u16;
+pub type CodeIndex = u8;
+
