+++ src/storage/segment.rs
@@ -7,8 +7,8 @@ use std::sync::Arc;
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct SegmentRecord {
     pub id: String,
-    pub quantized_indices: Vec<u16>,
-    pub qjl_bits: Vec<i8>,
+    pub quantized_indices: Vec<u8>,
+    pub qjl_bits: Vec<u8>,
     pub gamma: f32,
     #[serde(default)]
     pub is_deleted: bool,
@@ -219,4 +219,5 @@ impl SegmentManager {
             .collect()
     }
 }
+
